# YJIT: Lazily enable YJIT after prelude