--- a/internal/cmdlineopt.h
+++ b/internal/cmdlineopt.h
@@ -26,6 +26,9 @@ typedef struct ruby_cmdline_options {
 #if USE_MJIT
     struct mjit_options mjit;
 #endif
+#if USE_YJIT
+    bool yjit;
+#endif
 
     int sflag, xflag;
     unsigned int warning: 1;