Currently, YJIT compiles code for prelude as well. But some code used by prelude may not be used ever again after prelude, which could bring less localized code and unnecessary pressure on code GC.

JITing prelude seems just useless if it doesn't contribute to speeding up prelude. And the cost of JITing prelude seems to outweigh the JITed code's speedup. So I'd like to propose not JITing prelude.

### Before
```
$ time ruby -v --yjit -e ''
ruby 3.2.0dev (2022-10-20T00:45:36Z master 2e3b764ad1) +YJIT [arm64-darwin21]
ruby -v --yjit -e ''  0.05s user 0.02s system 91% cpu 0.071 total
```

### After
```
$ time ruby -v --yjit -e ''
ruby 3.2.0dev (2022-10-20T00:45:54Z yjit-lazy-init f60a1b4ce3) +YJIT [arm64-darwin21]
ruby -v --yjit -e ''  0.04s user 0.01s system 90% cpu 0.058 total
```

I've tried both multiple times, and this PR was reliably faster.