@noahgibbs 2022-10-20
This looks like a good idea to me.