@maximecb 2022-10-24
Just to make sure I understand, this enables YJIT after Ruby is done loading Ruby runtime code? Is that what you mean by prelude?