@noahgibbs 2022-10-24
The prelude is the pre-loaded things like Rubygems which are loaded before the "real start" of the Ruby code in the file you specify (or equivalent -e param(s) to eval). So this won't affect e.g. Rails initialisers, they happen long after the prelude. But it will keep us from starting to JIT Rubygems' startup code, which seems like a good idea to me.

It would probably be an even better idea to have some way to not start JITting until an app-specified location, e.g. after Rails startup. But that would need more of an API. This has the advantage that it's really clear where to start.